[IMP] account,l10n_de: move audit log attachment tests
In this commit we move the attachment audit trail tests
from l10n_de to account. In odoo#203229, the attachment audit
trail checks were moved from to account, but the tests remained
in the localization. An additional case covering the attachments
in log notes was also added to the tests.

opw-5107912
